Modules/requirements.txt
requests==2.28.2
GitPython==3.1.31
jinja2==3.1.2
pandas==1.5.3
pika==1.3.1
python-dotenv==1.0.0
SQLAlchemy==2.0.4
mysqlclient==2.1.1
smartsheet-python-sdk==3.0.0
openpyxl==3.1.1